Fix playback of Tunein radio streams

+++ music_assistant/helpers/database.py
@@ -11,7 +11,7 @@ from music_assistant.helpers.typing import MusicAssistant
 
 # pylint: disable=invalid-name
 
-SCHEMA_VERSION = 1
+SCHEMA_VERSION = 2
 
 
 class Database:
@@ -141,19 +141,26 @@ class Database:
                 SCHEMA_VERSION,
             )
 
-            # schema version 1: too many breaking changes, simply drop the media tables for now
-            async with self.get_db() as _db:
-                await _db.execute("DROP TABLE IF EXISTS artists")
-                await _db.execute("DROP TABLE IF EXISTS albums")
-                await _db.execute("DROP TABLE IF EXISTS tracks")
-                await _db.execute("DROP TABLE IF EXISTS playlists")
-                await _db.execute("DROP TABLE IF EXISTS radios")
-                await _db.execute("DROP TABLE IF EXISTS playlist_tracks")
-                await _db.execute("DROP TABLE IF EXISTS album_tracks")
-                await _db.execute("DROP TABLE IF EXISTS provider_mappings")
-                await _db.execute("DROP TABLE IF EXISTS cache")
+            if prev_version < 1:
+                # schema version 1: too many breaking changes, simply drop the media tables for now
+                async with self.get_db() as _db:
+                    await _db.execute("DROP TABLE IF EXISTS artists")
+                    await _db.execute("DROP TABLE IF EXISTS albums")
+                    await _db.execute("DROP TABLE IF EXISTS tracks")
+                    await _db.execute("DROP TABLE IF EXISTS playlists")
+                    await _db.execute("DROP TABLE IF EXISTS radios")
+                    await _db.execute("DROP TABLE IF EXISTS playlist_tracks")
+                    await _db.execute("DROP TABLE IF EXISTS album_tracks")
+                    await _db.execute("DROP TABLE IF EXISTS provider_mappings")
+                    await _db.execute("DROP TABLE IF EXISTS cache")
 
-            # store current schema version
-            await self.insert_or_replace(
-                "settings", {"key": "version", "value": str(SCHEMA_VERSION)}
-            )
+            if prev_version < 2:
+                # schema version 2: repair invalid data for radio items
+                async with self.get_db() as _db:
+                    await _db.execute("DROP TABLE IF EXISTS radios")
+                    await self.delete("provider_mappings", {"media_type": "radio"}, _db)
+
+        # store current schema version
+        await self.insert_or_replace(
+            "settings", {"key": "version", "value": str(SCHEMA_VERSION)}
+        )
